Page와 Page 테스트를 제거한다.
App컴포넌트와 Page컴포넌트ㅑ가 거의 동일해져서 Page컴포넌트를 지우고
App컴포넌트에 통합한다.Page와 Page 테스트를 제거한다.
